Rename shared SQLModel package
